+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove `bootstrap-nav-walker` module
 * Remove `remove-default-image-sizes` module
 * Rename `remove-update-notive` to `disable-update-notice`
+* Rename `disable-customization` to `disable-customizer`
 
 ### 2.1.0: Gen 9th, 2018
 * Update `itsensoul` organization name with `helloensoul`